requirements.txt
# Spotify AI DJ - Python dependencies
# Install with:  pip install -r requirements.txt
# Windows users: double-click install.bat instead
# Linux users:   run bash install.sh instead
# macOS users:   run bash install_mac.sh instead

# Required on all platforms
customtkinter>=5.2.0    # GUI for Windows, macOS, and Linux X11
spotipy>=2.23.0         # Spotify Web API client
google-genai>=1.0.0     # Google Gemini AI client
pydantic>=2.0.0         # Structured AI response parsing

# GTK4 backend (Linux Wayland only) - NOT installed via pip
# The install.sh script handles this automatically.
# To install manually:
#   Debian/Ubuntu:  sudo apt install python3-gi python3-gi-cairo gir1.2-gtk-4.0
#   Arch:           sudo pacman -S python-gobject gtk4
#   Fedora:         sudo dnf install python3-gobject gtk4